+++ Sources/VCore/Extensions/SwiftUI/View/View.GetFrame.swift
@@ -9,6 +9,33 @@ import SwiftUI
 
 // MARK: - View Get Frame
 extension View {
+    /// Retrieves bounds from `View`.
+    ///
+    ///     @State private var frame: CGRect = .init()
+    ///
+    ///     var body: some View {
+    ///         Color.accentColor
+    ///             .getFrame(in: .global, { frame = $0 })
+    ///     }
+    ///
+    @available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
+    public func getFrame(
+        in coordinateSpace: some CoordinateSpaceProtocol,
+        _ action: @escaping (CGRect) -> Void
+    ) -> some View {
+        self
+            .background(content: {
+                GeometryReader(content: { proxy in
+                    Color.clear
+                        .preference(
+                            key: FramePreferenceKey.self,
+                            value: proxy.frame(in: coordinateSpace)
+                        )
+                        .onPreferenceChange(FramePreferenceKey.self, perform: action)
+                })
+            })
+    }
+
     /// Retrieves `CGRect` from `View`.
     ///
     ///     @State private var frame: CGRect = .init()
@@ -18,7 +45,7 @@ extension View {
     ///             .getFrame(in: .global, { frame = $0 })
     ///     }
     ///
-    public func getFrame(
+    public func getFrame( // TODO: iOS 17.0 - Remove
         in coordinateSpace: CoordinateSpace,
         _ action: @escaping (CGRect) -> Void
     ) -> some View {